Give the Tau-BA factoring helpers one declaration site
satisfiability.h forward-declared factored_tau_sat/factored_tau_valid as
static templates defined in tau_ba.tmpl.h, which only matched because the
amalgamated tau.h is a single translation unit (review nit from #91).
Move the option, ba_component_factoring_enabled() and the declarations of
the three helpers into boolean_algebras/tau_ba_factoring.h, included by
both tau_ba.h and satisfiability.h, and drop static from the template
definitions. No behavioural change.
